Package scrape code outside of CLI
