+++ go.mod
@@ -8,7 +8,10 @@ require (
 	github.com/gobwas/glob v0.2.3 // indirect
 	github.com/gocolly/colly v1.1.0
 	github.com/kennygrant/sanitize v1.2.4 // indirect
+	github.com/mitchellh/go-homedir v1.0.0
 	github.com/saintfish/chardet v0.0.0-20120816061221-3af4cd4741ca // indirect
+	github.com/spf13/cobra v0.0.3
+	github.com/spf13/viper v1.3.1
 	github.com/temoto/robotstxt v0.0.0-20180810133444-97ee4a9ee6ea // indirect
 	golang.org/x/net v0.0.0-20181220203305-927f97764cc3 // indirect
 	google.golang.org/appengine v1.4.0 // indirect
